add casts to eeprom

+++ src/eeprom.c
@@ -12,7 +12,7 @@ void eeprom_write(SIZE_T base_address, U8_T* p_buf, SIZE_T len)
 {
     SIZE_T i;
 
-    for (i = 0u; i < len; ++i) {
+    for (i = (SIZE_T) 0u; i < len; ++i) {
         eeprom_write_byte((SIZE_T)( base_address + i ), p_buf[i] );
     }
 }
@@ -32,7 +32,7 @@ void eeprom_read(SIZE_T base_address, U8_T* p_buf, SIZE_T len)
 {
     SIZE_T i;
 
-    for (i = 0u; i < len; ++i) {
+    for (i = (SIZE_T) 0u; i < len; ++i) {
         p_buf[i] = eeprom_read_byte((SIZE_T)( base_address + i ));
     }
 }
@@ -48,7 +48,7 @@ void eeprom_erase(SIZE_T base_address, SIZE_T len)
 {
     SIZE_T i;
 
-    for (i = 0u; i < len; ++i) {
+    for (i = (SIZE_T) 0u; i < len; ++i) {
         eeprom_erase_byte((SIZE_T)( base_address + i ));
     }
 }
